// src/state/types.ts
export type AppThunk<ReturnType = void> = ThunkAction<
  ReturnType,
  State,
  unknown,
  AnyAction
>;
// ...
export interface BlockState {
  currentBlock: number;
  initialBlock: number;
}
// ...
export interface TokenPrices {
  symbol: string;
  address: Address;
  price: number;
  decimals: number;
}

export interface LpTokenPrices {
  symbol: string;
  pid: number;
  address: Address;
  price: number;
  decimals: number;
}

// Slices states

export interface ToastsState {
  data: Toast[];
}

export interface FarmsState {
  data: Farm[];
}

export interface PoolsState {
  data: Pool[];
}

export interface NfaState {
  isInitialized: boolean;
  isLoading: boolean;
  data: Nfa[];
}

export interface DualFarmsState {
  data: DualFarm[];
}

export interface JungleFarmsState {
  data: JungleFarm[];
}

export interface BillsState {
  data: Bills[];
}

export interface NetworkState {
  isInitialized: boolean;
  isLoading: boolean;
  data: Network;
}

export interface VaultsState {
  loadVaultData: boolean;
  userDataLoaded: boolean;
  data: Vault[];
}

export interface NfaStakingPoolsState {
  data: NfaStakingPool[];
}

export interface ProfileState {
  isInitialized: boolean;
  isLoading: boolean;
  data: Profile;
}

export interface TagsType {
  [key: string]: any;
}

export interface OrderingType {
  [key: string]: any;
}

export interface StatsState {
  isInitialized: boolean;
  isLoading: boolean;
  HomepageData: HomepageData;
  HomepageTokenStats: HomepageTokenStats[];
  HomepageNews: NewsCardType[];
  HomepageLaunchCalendar: LaunchCalendarCard[];
  HomepageServiceStats: ServiceData[];
  FarmLpAprs: FarmLpAprsType;
  Tags: TagsType;
  Ordering: OrderingType;
  LiveIfo: LiveIfo[];
  data: Stats;
}

export interface AuctionsState {
  isInitialized: boolean;
  isLoading: boolean;
  data: AuctionsOverall;
}

export interface TokenPricesState {
  isTokensInitialized: boolean;
  isInitialized: boolean;
  isLoading: boolean;
  tokens: Token[];
  digichainPrice: string;
  data: TokenPrices[];
}
export interface LpTokenPricesState {
  isInitialized: boolean;
  isLoading: boolean;
  data: LpTokenPrices[];
}

export interface StatsOverallState {
  isInitialized: boolean;
  isLoading: boolean;
  data: StatsOverall;
}

export interface IazosState {
  isInitialized: boolean;
  isLoading: boolean;
  iazoData: Iazo[];
  iazoDefaultSettings: IazoDefaultSettings;
}

export type TeamResponse = {
  0: string;
  1: string;
  2: string;
  3: string;
  4: boolean;
};

export type TeamsById = {
  [key: string]: Team;
};

export interface TeamsState {
  isInitialized: boolean;
  isLoading: boolean;
  data: TeamsById;
}

// Global state

export interface State {
  farms: FarmsState;
  block: BlockState;
  toasts: ToastsState;
  pools: PoolsState;
  profile: ProfileState;
  stats: StatsState;
  statsOverall: StatsOverallState;
  teams: TeamsState;
  auctions: AuctionsState;
  vaults: VaultsState;
  tokenPrices: TokenPricesState;
  lpTokenPrices: LpTokenPricesState;
  iazos: IazosState;
  network: NetworkState;
  nfaStakingPools: NfaStakingPoolsState;
  dualFarms: DualFarmsState;
  jungleFarms: JungleFarmsState;
  bills: BillsState;
  nfas: NfaState;
  protocolDashboard: ProtocolDashboardState;
  userStats: ApiResponse;
}